[AC-1495] Extract UpgradePlanAsync into a command (#3081)
* This is a pure lift & shift with no refactors

* Only register subscription commands in Api

---------
