+++ src/Admin/Startup.cs
@@ -9,8 +9,6 @@ using Stripe;
 using Microsoft.AspNetCore.Mvc.Razor;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Bit.Admin.Services;
-using Bit.Core.Repositories.Noop;
-using Bit.Core.SecretsManager.Repositories;
 
 #if !OSS
 using Bit.Commercial.Core.Utilities;
@@ -88,7 +86,6 @@ public class Startup
         services.AddBaseServices(globalSettings);
         services.AddDefaultServices(globalSettings);
         services.AddScoped<IAccessControlService, AccessControlService>();
-        services.AddScoped<IServiceAccountRepository, NoopServiceAccountRepository>();
 
 #if OSS
         services.AddOosServices();
