+++ src/Core/OrganizationFeatures/OrganizationSubscriptions/Interface/IUpgradeOrganizationPlanCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace Bit.Core.OrganizationFeatures.OrganizationSubscriptions.Interface;
+
+public interface IUpgradeOrganizationPlanCommand
+{
+    Task<Tuple<bool, string>> UpgradePlanAsync(Guid organizationId, Models.Business.OrganizationUpgrade upgrade);
+}
