+++ src/Identity/Startup.cs
@@ -4,8 +4,6 @@ using AspNetCoreRateLimit;
 using Bit.Core;
 using Bit.Core.Auth.Models.Business.Tokenables;
 using Bit.Core.Context;
-using Bit.Core.Repositories.Noop;
-using Bit.Core.SecretsManager.Repositories;
 using Bit.Core.Settings;
 using Bit.Core.Utilities;
 using Bit.Identity.Utilities;
@@ -47,8 +45,6 @@ public class Startup
 
         // Repositories
         services.AddDatabaseRepositories(globalSettings);
-
-        services.AddScoped<IServiceAccountRepository, NoopServiceAccountRepository>();
 
         // Context
         services.AddScoped<ICurrentContext, CurrentContext>();
@@ -154,7 +150,6 @@ public class Startup
             client.BaseAddress = new Uri(globalSettings.BaseServiceUri.InternalSso);
         });
     }
-
 
     public void Configure(
         IApplicationBuilder app,
